add normalize in filter, better view for all elements in paginator, add side for tooltip, add custom field in table component and add getNestedValue, change table in invoices por table component,update backend basic for send token in headers

+++ src/services/base.service.ts
@@ -2,12 +2,14 @@
 import axios from 'axios';
 import { BaseResponseLogin, BaseResponse } from './base.interface';
 
+const getToken = localStorage.getItem('token');
+
 export const api = axios.create({
     baseURL: `${import.meta.env.VITE_BASE_URL_API}/api`
 })
 
 export const getDataApi = (endpoint: string) => {
-    return api.get(endpoint).then((response) => {
+    return api.get(endpoint, { headers: { Authorization: getToken } }).then((response) => {
         return response.data;
     }).catch(err => {
         return err.response.data;
@@ -16,8 +18,9 @@ export const getDataApi = (endpoint: string) => {
 
 export const getDataFileApi = (endpoint: string) => {
     return api.get(endpoint, {
-        responseType: 'blob'
-    }).then((response) => {
+        responseType: 'blob',
+        headers: { Authorization: getToken }
+    },).then((response) => {
         return response.data;
     }).catch(err => {
         return err.response.data;
@@ -42,7 +45,7 @@ export const postDataFileGetApi = async (endpoint: string, data: any) => {
 }
 
 export const postDataApi = async (endpoint: string, data: any): Promise<BaseResponseLogin | BaseResponse | any> => {
-    return await api.post(endpoint, data).then((response) => {
+    return await api.post(endpoint, data, { headers: { Authorization: getToken } }).then((response) => {
         return response.data;
     }).catch((err) => {
         return err.response.data;
@@ -53,6 +56,7 @@ export const postFilesDataApi = async (endpoint: string, formData: FormData): Pr
     return await api.put(endpoint, formData, {
         headers: {
             "Content-Type": "multipart/form-data",
+            Authorization: getToken
         },
     }).then((response) => {
         return response.data;
@@ -69,7 +73,7 @@ export const postDataFileApi = async (endpoint: string, data: any): Promise<Base
 }
 
 export const putDataApi = async (endpoint: string, data: any): Promise<BaseResponse> => {
-    return await api.put(endpoint, data).then((response) => {
+    return await api.put(endpoint, data, { headers: { Authorization: getToken } }).then((response) => {
         return response.data;
     }).catch((err) => {
         return err.response.data;
@@ -77,7 +81,7 @@ export const putDataApi = async (endpoint: string, data: any): Promise<BaseRespo
 }
 
 export const deleteDataApi = async (endpoint: string): Promise<BaseResponse> => {
-    return await api.delete(endpoint).then((response) => {
+    return await api.delete(endpoint, { headers: { Authorization: getToken } }).then((response) => {
         return response.data;
     }).catch((err) => {
         return err.response.data;
